Moved database functions and config to seperate files
