+++ firebase.config.js
@@ -1,0 +1,12 @@
+const firebaseConfig = {
+    apiKey: "AIzaSyC-RiwBX8Si5IKxjIt7GgVTy7TNxf-hq4E",
+    authDomain: "human-movement-mapping.firebaseapp.com",
+    databaseURL: "https://human-movement-mapping-default-rtdb.firebaseio.com",
+    projectId: "human-movement-mapping",
+    storageBucket: "human-movement-mapping.appspot.com",
+    messagingSenderId: "262748753205",
+    appId: "1:262748753205:web:22aff0982462db9ccb907c",
+    measurementId: "G-QB5JMW811H"
+  };
+
+export default firebaseConfig;
